requirements-test.txt
requests-mock==1.3.0
coverage==4.4.1
flake8==3.3.0
pep257==0.7.0
doc8==0.7.0
Pygments==2.2.0
mypy==0.511
